split proof logic into multihop/multihop_helper packages to break cyclic dependencies
