+++ modules/core/04-channel/keeper/handshake.go
@@ -14,6 +14,7 @@ import (
 	porttypes "github.com/cosmos/ibc-go/v7/modules/core/05-port/types"
 	host "github.com/cosmos/ibc-go/v7/modules/core/24-host"
 	"github.com/cosmos/ibc-go/v7/modules/core/exported"
+	mh "github.com/cosmos/ibc-go/v7/modules/core/multihop"
 )
 
 // ChanOpenInit is called by a module to initiate a channel opening handshake with
@@ -171,11 +172,28 @@ func (k Keeper) ChanOpenTry(
 		counterpartyHops, counterpartyVersion,
 	)
 
-	if err := k.connectionKeeper.VerifyChannelState(
-		ctx, connectionEnd, proofHeight, proofInit,
-		counterparty.PortId, counterparty.ChannelId, expectedChannel,
-	); err != nil {
-		return "", nil, err
+	if len(connectionHops) > 1 {
+		clientID := "client-0"
+		clientState, found := k.clientKeeper.GetClientState(ctx, clientID)
+		if !found {
+			return "", nil, sdkerrors.Wrapf(clienttypes.ErrClientNotFound, "client state not found for client id: %s", clientID)
+		}
+		consensusState, found := k.clientKeeper.GetClientConsensusState(ctx, clientID, clientState.GetLatestHeight())
+		if !found {
+			return "", nil, sdkerrors.Wrapf(clienttypes.ErrConsensusStateNotFound, "consensus state %s not found for client id: %s", clientID)
+		}
+		conStateProof := make([]*mh.ConsStateProof, 3)
+		val := []byte("test")
+		if err := mh.VerifyMultiHopProofMembership(consensusState, clientState, k.cdc, conStateProof, val); err != nil {
+			return "", nil, err
+		}
+	} else {
+		if err := k.connectionKeeper.VerifyChannelState(
+			ctx, connectionEnd, proofHeight, proofInit,
+			counterparty.PortId, counterparty.ChannelId, expectedChannel,
+		); err != nil {
+			return "", nil, err
+		}
 	}
 
 	var (
